+++ src/mineserver/network/message/spawnposition.h
@@ -35,9 +35,9 @@ namespace Mineserver
 {
   struct Network_Message_SpawnPosition : public Mineserver::Network_Message
   {
-    int16_t x;
-    int16_t y;
-    int16_t z;
+    int32_t x;
+    int32_t y;
+    int32_t z;
   };
 }
